+++ R/amRangeAreaChart.R
@@ -16,6 +16,14 @@
 #'   \code{name} for the legend label, \code{color} for the color of the range
 #'   area, and \code{opacity} for the opacity of the range area, a number
 #'   between 0 and 1
+#' @param hline an optional horizontal line to add to the chart; it must be a
+#'   named list of the form \code{list(value = h, line = settings)} where
+#'   \code{h} is the "intercept" and \code{settings} is a list of settings
+#'   created with \code{\link{amLine}}
+#' @param vline an optional vertical line to add to the chart; it must be a
+#'   named list of the form \code{list(value = v, line = settings)} where
+#'   \code{v} is the "intercept" and \code{settings} is a list of settings
+#'   created with \code{\link{amLine}}
 #' @param xLimits range of the x-axis, a vector of two values specifying
 #'   the left and right limits of the x-axis; \code{NULL} for default values
 #' @param yLimits range of the y-axis, a vector of two values specifying
@@ -249,6 +257,8 @@ amRangeAreaChart <- function(
   xValue,
   yValues,
   areas = NULL, # default
+  hline = NULL,
+  vline = NULL,
   xLimits = NULL,
   yLimits = NULL,
   expandX = 0,
@@ -657,6 +667,22 @@ amRangeAreaChart <- function(
     chartId <- paste0("rangeareachart-", randomString(15))
   }
 
+  if(!is.null(hline)){
+    if(any(!is.element(c("value", "line"), names(hline)))){
+      stop(
+        "Invalid `hline` argument."
+      )
+    }
+  }
+
+  if(!is.null(vline)){
+    if(any(!is.element(c("value", "line"), names(vline)))){
+      stop(
+        "Invalid `vline` argument."
+      )
+    }
+  }
+
   # describe a React component to send to the browser for rendering.
   component <- reactR::component(
     "AmRangeAreaChart",
@@ -671,6 +697,8 @@ amRangeAreaChart <- function(
       maxX = xLimits[2L],
       minY = yLimits[1L],
       maxY = yLimits[2L],
+      hline = hline,
+      vline = vline,
       chartTitle = chartTitle,
       theme = theme,
       draggable = draggable,